metodo actualizar - ver xq cuando actualiza se setea en null el paciente.

+++ src/main/java/com/example/Odontologica2/model/TurnoDTO.java
@@ -1,11 +1,9 @@
 package com.example.Odontologica2.model;
 
-import com.example.Odontologica2.persistence.entities.Odontologo;
-
 import java.util.Date;
 
 public class TurnoDTO {
-    private Integer id;
+    private Long id;
     private Date date;
     private PacienteDTO pacienteDTO;
     private OdontologoDTO odontologoDTO;
@@ -26,11 +24,11 @@ public class TurnoDTO {
         this.odontologoDTO = odontologoDTO;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
